Replacing custom unique id with UUID v1
Previous implementation relied on undocumented custom function to generate
unique identifyers. I have replaced it, with python UUID v1 built in,
as it provides the same functionality in a safer way.

UUIDv1 is documented in the RFC4122[0], along with rest of the UUID standard.
The node MAC address is retrieved using python built-in `uuid.get_node()` function,
which has already been used by the existing unique identifier function.

[0] https://www.rfc-editor.org/rfc/rfc4122.html
[1] https://docs.python.org/3.8/library/uuid.html?highlight=uuid#uuid.uuid1

Signed-off-by: Jiri Podivin <jpodivin@redhat.com>